+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Review permissions, distinguishing client-only, common and server-only ones, and no more dedicated entities/records
+    - Have a single autorun for each collection dependency on package configuration
 
 ### 1.6.0